# perf(collections): use `for` loop instead of `forEach()` in `unzip()`