This change makes `unzip()` approximately 50%-61% faster.

<details open>
<summary>Benchmark</summary>

```ps1
PS ~\console> deno bench mod.ts
Check file:///D:/coding/project/javascript/deno/console/mod.ts
cpu: Intel(R) Core(TM) i3-6006U CPU @ 2.00GHz
runtime: deno 1.44.4 (x86_64-pc-windows-msvc)

file:///D:/coding/project/javascript/deno/console/mod.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------
before        625.71 ns/iter   1,598,189.9   (586.87 ns … 1.12 µs) 627.6 ns 1.12 µs 1.12 µs
after         243.75 ns/iter   4,102,578.2   (210.64 ns … 1.03 µs) 234.33 ns 478.63 ns 549.52 ns
```
</details>

<details>
<summary>Code</summary>

```ts
import { unzip as unzip_before } from "@std/collections";

function unzip_after<T, U>(pairs: readonly [T, U][]): [T[], U[]] {
  const { length } = pairs;

  const result: [T[], U[]] = [
    new Array<T>(length),
    new Array<U>(length),
  ];

  for (let i = 0; i < length; ++i) {
    const pair = pairs[i] as [T, U];

    result[0][i] = pair[0];
    result[1][i] = pair[1];
  }

  return result;
}

const t: [number, string][] = [
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
  [1, "a"], [2, "b"], [3, "c"], [4, "d"],
];

Deno.bench("before", () => {
  unzip_before(t);
});


Deno.bench("after", () => {
  unzip_after(t);
});

```
</details>