@babiabeo 2024-06-21
> Can you share the benchmark code?

It was already in the description 👍